Move `SyncedParametersList` to qb
